Crate wt_flight_recorder

Flight Data Recorder

This module allows for easily recording important flight information for later analysis and debugging.

Data is output to a file in the \work directory with the initialization time as a prefix and .msgpack.gz as the extension. Files produced are serialized as MsgPack and then processed by a gzip encoder. Over the course of a session, a log may be broken across multiple files in order to keep the size of each individual log file reasonable. Files belonging to the same session will have the same prefix, but have an incrementing number associated with them.

These files can then be later processed by the wt_flight_to_csv processor to convert the data into a CSV format, suitable for further analysis.

use wt_flight_recorder::FlightDataRecorder;

#[derive(serde::Serialize, serde::Deserialize)]
struct MyData {
    time: f64,
    thrust: f64,
    output: f64,
}

let mut recorder: FlightDataRecorder<MyData> = FlightDataRecorder::new().unwrap();

let publish_result = recorder.publish(&MyData {
    time: 0.0,
    thrust: 1000.0,
    output: 0.9,
});

if let Err(e) = publish_result {
    eprintln!("Unable to log event: {}", e);
}

Structs

FlightDataRecorder
A flight data recorder for aircraft data